+++ packages/cryptoplease/lib/app/screens/authenticated/receive_flow/routes.dart
@@ -1,10 +1,7 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/screens/authenticated/receive_flow/share_address_screen.dart';
 import 'package:cryptoplease/features/payment_request/presentation/link_details/routes.dart';
 import 'package:cryptoplease/features/payment_request/presentation/link_request/routes.dart';
 
 const receiveFlowRoutes = [
-  AutoRoute<void>(page: ShareAddressScreen),
   linkRequestRoutes,
   ...linkDetailsRoutes,
 ];