+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/app/screens/authenticated/activities/activities_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/authenticated_flow_screen.dart';
+import 'package:cryptoplease/app/screens/authenticated/cash_out/cash_out_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/home_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/investments/routes.dart';
 import 'package:cryptoplease/app/screens/authenticated/profile/edit_profile_screen.dart';
@@ -9,7 +10,7 @@ import 'package:cryptoplease/app/screens/authenticated/profile/menu_screen.dart'
 import 'package:cryptoplease/app/screens/authenticated/profile/profile_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/receive_flow/routes.dart';
 import 'package:cryptoplease/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart';
-import 'package:cryptoplease/features/add_funds/presentation/routes.dart';
+import 'package:cryptoplease/features/add_funds/presentation/add_funds_screen.dart';
 import 'package:cryptoplease/features/app_lock/presentation/routes.dart';
 import 'package:cryptoplease/features/backup_phrase/presentation/routes.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/presentation/routes.dart';
@@ -42,7 +43,8 @@ const authenticatedFlowRoutes = AutoRoute<dynamic>(
     ...iskpRoutes,
     AutoRoute<QrScannerRequest>(page: QrScannerScreen),
     ...receiveFlowRoutes,
-    ...addFundsRoutes,
+    AutoRoute<void>(page: AddFundsScreen, fullscreenDialog: true),
+    AutoRoute<void>(page: CashOutScreen, fullscreenDialog: true),
     appLockSetupFlowRoutes,
     AutoRoute<void>(page: ProfileScreen, fullscreenDialog: true),
     AutoRoute<void>(page: EditProfileScreen),