+++ packages/cryptoplease/lib/features/payment_request/presentation/link_details/flow.dart
@@ -1,15 +1,16 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/components/tx_result_screen.dart';
 import 'package:cryptoplease/app/routes.dart';
 import 'package:cryptoplease/core/presentation/format_amount.dart';
 import 'package:cryptoplease/di.dart';
 import 'package:cryptoplease/features/payment_request/bl/payment_request.dart';
 import 'package:cryptoplease/features/payment_request/bl/payment_request_verifier/bloc.dart';
 import 'package:cryptoplease/features/payment_request/bl/repository.dart';
+import 'package:cryptoplease/features/payment_request/presentation/link_details/tx_result_screen.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
 import 'package:cryptoplease/l10n/device_locale.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/app_bar.dart';
+import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
@@ -41,25 +42,27 @@ class _LinkDetailsFlowScreenState extends State<LinkDetailsFlowScreen> {
   }
 
   @override
-  Widget build(BuildContext context) => StreamBuilder<PaymentRequest>(
-        stream: _stream,
-        builder: (context, snapshot) {
-          final data = snapshot.data;
-          if (data == null) return const _Loader();
+  Widget build(BuildContext context) => CpTheme.dark(
+        child: StreamBuilder<PaymentRequest>(
+          stream: _stream,
+          builder: (context, snapshot) {
+            final data = snapshot.data;
+            if (data == null) return const _Loader();
 
-          return data.state.when(
-            initial: () => BlocProvider<PaymentRequestVerifierBloc>(
-              create: (_) => sl<PaymentRequestVerifierBloc>(param1: data),
-              lazy: false,
-              child: Provider<PaymentRequest>.value(
-                value: data,
-                child: const AutoRouter(),
+            return data.state.when(
+              initial: () => BlocProvider<PaymentRequestVerifierBloc>(
+                create: (_) => sl<PaymentRequestVerifierBloc>(param1: data),
+                lazy: false,
+                child: Provider<PaymentRequest>.value(
+                  value: data,
+                  child: const AutoRouter(),
+                ),
               ),
-            ),
-            completed: (_) => _Success(request: data),
-            failure: () => _Failure(request: data),
-          );
-        },
+              completed: (_) => _Success(request: data),
+              failure: () => _Failure(request: data),
+            );
+          },
+        ),
       );
 }