+++ packages/cryptoplease/lib/features/payment_request/presentation/link_details/share_link.dart
@@ -1,6 +1,4 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/components/share_message/header.dart';
-import 'package:cryptoplease/app/components/share_message_wrapper.dart';
 import 'package:cryptoplease/core/presentation/format_amount.dart';
 import 'package:cryptoplease/core/presentation/utils.dart';
 import 'package:cryptoplease/core/tokens/token_list.dart';
@@ -11,6 +9,8 @@ import 'package:cryptoplease/ui/app_bar.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/colors.dart';
 import 'package:cryptoplease/ui/content_padding.dart';
+import 'package:cryptoplease/ui/share_message/share_message_bubble.dart';
+import 'package:cryptoplease/ui/share_message/share_message_header.dart';
 import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -50,14 +50,12 @@ class SharePaymentRequestLinkScreen extends StatelessWidget {
       onPressed: () => Share.share(message),
     );
 
-    final messageWrapper = ShareMessageWrapper(
+    final messageBubble = ShareMessageBubble(
       textSpan: TextSpan(
         children: [
-          WidgetSpan(
-            child: ShareMessageHeader(
-              intro: context.l10n.sharePaymentRequestLinkIntro,
-              amount: amount,
-            ),
+          ShareMessageHeader(
+            intro: context.l10n.sharePaymentRequestLinkIntro,
+            amount: amount,
           ),
           const WidgetSpan(child: _Instructions()),
           WidgetSpan(child: _Links(link: request.dynamicLink)),
@@ -81,7 +79,7 @@ class SharePaymentRequestLinkScreen extends StatelessWidget {
                 ),
                 child: subtitle,
               ),
-              Flexible(child: messageWrapper),
+              Flexible(child: messageBubble),
               const SizedBox(height: 24),
               shareButton,
             ],